requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==65.0.0
django-cors-headers==4.4.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
idna==3.10
psycopg2==2.9.9
python-decouple==3.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
